updated examples for 0.2.0
