--- XRTK.Examples/Assets/XRTK.Examples/Inspectors.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a0ba587a84294a0e9e77d9a561fb2b70
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
