--- XRTK.Examples/Assets/XRTK.Examples/Inspectors/DemoCustomExtensionServiceProfileInspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d070a958693f4d59ab3f6d6619390373
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
