--- XRTK.Examples/Assets/XRTK.Examples/Inspectors/MaterialMatrixInspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 95a1b33dab6d4005950b0877d3f91046
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
