--- XRTK.Examples/Assets/XRTK.Examples/Inspectors/XRTK.Examples.Inspectors.asmdef
@@ -1,18 +1,0 @@
-{
-    "name": "XRTK.Examples.Inspectors",
-    "references": [
-        "XRTK",
-        "XRTK.Examples",
-        "XRTK.Inspectors"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": []
-}
