--- XRTK.Examples/Assets/XRTK.Examples/Inspectors/XRTK.Examples.Inspectors.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 55f07d1b464e41cab0bb020831cde2e8
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
